requirements.txt
torch
torchvision
filterpy==1.4.5
dlib==19.20.0
mss==6.0.0
numpy==1.18.5
requests==2.24.0
Eel==0.13.2
opencv_python
ewmh==0.1.6
scipy==1.5.1
lap==0.4.0
numba==0.50.1
pandas==1.0.5
Pillow==7.2.0
PyQt5==5.15.0
tensorflow==2.2.0
tqdm==4.47.0
Xlib==0.21
